+++ packages/client/src/components/SafeDetails.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Dropdown from "./Dropdown";
+import { Dropdown } from "library/components";
 import { SAFE_TYPES_TO_META } from "constants/maps";
 
 const safeTypes = Object.entries(SAFE_TYPES_TO_META).map((type) => ({
