+++ packages/client/src/components/SafeSettings/components/EditThreshold/EditThresholdForm.js
@@ -1,6 +1,6 @@
 import { getProgressPercentageForSignersAmount } from "utils";
 import Svg from "library/Svg";
-import ProgressBar from "components/ProgressBar";
+import { ProgressBar } from "library/components";
 
 const EditThresholdForm = ({
   newThreshold,
