+++ packages/client/src/components/index.js
@@ -2,8 +2,6 @@ export { default as Navigation } from "./Navigation";
 export { default as Transactions } from "./Transactions";
 export { default as Logo } from "./Logo";
 export { default as WalletPrompt } from "./WalletPrompt";
-export { default as Dropdown } from "./Dropdown";
-export { default as ProgressBar } from "./ProgressBar";
 export { default as Loading } from "./Loading";
 export { default as SafeDetails } from "./SafeDetails";
 export { default as SafeOwners } from "./SafeOwners";
